requirements-dev.txt
pytest
pytest-django
pytest-cov
pandas
bandit
pip-audit
selenium